pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "voice-agent"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "websockets>=14.0",
    "python-dotenv>=1.0.0",
    "langchain>=0.3.0",
    "langchain-anthropic>=0.3.0",
    "langchain-core>=0.3.0",
    "langgraph>=0.2.0",
    "langgraph-checkpoint-postgres>=2.0.0",
    "langgraph-checkpoint-redis>=1.0.0",
    "soundfile>=0.13.0",
    "numpy>=1.24.0",
    "torch>=2.0.0",
    "torchaudio>=2.0.0",
    "chatterbox-tts @ git+https://github.com/resemble-ai/chatterbox.git",
    # SDR Agent dependencies
    "twilio>=9.0.0",
    "selenium>=4.15.0",
    "webdriver-manager>=4.0.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
    "click>=8.1.0",
    "rich>=13.0.0",
    "httpx>=0.27.0",
    "pydantic>=2.5.0",
    "phonenumbers>=8.13.0",
    "scipy>=1.11.0",
    "pytz>=2024.1",
]

[project.scripts]
sdr = "sdr_agent.cli:cli"

[dependency-groups]
dev = [
    "ruff>=0.8.0",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/sdr_agent"]

[tool.uv]
override-dependencies = [
    "numpy>=1.26.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"
